requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pandas==2.2.0
numpy==1.26.0

